moved SidebarItem into its own component and fixed imports
